requirements.txt
gradio 
tensorflow 
tensorflow_hub 
numpy 
pillow